+++ carp_mobile_sensing/lib/runtime/runtime.dart
@@ -17,6 +17,8 @@ import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:permission_handler_platform_interface/permission_handler_platform_interface.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:uuid/uuid.dart';
 
 part 'probes.dart';
 part 'probe_registry.dart';
@@ -26,6 +28,7 @@ part 'probe_controller.dart';
 part 'sampling_package.dart';
 part 'data_manager.dart';
 part 'permission_handler.dart';
+part 'settings.dart';
 
 /// Generic sensing exception.
 class SensingException implements Exception {