+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.8.0
+* using the Dart `Duration` class in many places where duration previously was specified using miliseconds.
+* added the `MarkedMeasure` class for creating measures that collect all measures since last time data was collected.
+Very useful for collecting things like calendar entries, log entries, and health data.
+
+
 ## 0.7.2
 * added support for better logging using the `DebugLevel` settings